Adjust poller to use same configuration parameters
