+++ test/utils.test.js
@@ -11,7 +11,7 @@ governing permissions and limitations under the License.
 */
 
 const { useMockServer } = require('./mock-server');
-const { errorResponse, stringParameters, checkMissingRequestInputs, getBearerToken, request, requestSpreadsheet, getConfig, requestSaaS } = require('./../actions/utils.js');
+const { errorResponse, stringParameters, checkMissingRequestInputs, getBearerToken, request, requestSpreadsheet, getConfig, requestSaaS, getProductUrl } = require('./../actions/utils.js');
 const { http, HttpResponse } = require('msw');
 
 test('interface', () => {
@@ -265,5 +265,32 @@ describe('request', () => {
     }));
 
     await expect(request('testRequest', 'https://example.com/timeout', {}, 100)).rejects.toThrow('This operation was aborted');
+  });
+});
+
+describe('getProductUrl', () => {
+  test('getProductUrl with urlKey and sku', () => {
+      const context = { storeUrl: 'https://example.com', pathFormat: '/products/{urlKey}/{sku}' };
+      expect(getProductUrl({ urlKey: 'my-url-key', sku: 'my-sku' }, context)).toBe('https://example.com/products/my-url-key/my-sku');
+  });
+
+  test('getProductUrl with urlKey', () => {
+      const context = { storeUrl: 'https://example.com', pathFormat: '/{urlKey}' };
+      expect(getProductUrl({ urlKey: 'my-url-key' }, context)).toBe('https://example.com/my-url-key');
+  });
+
+  test('return null for missing storeUrl', () => {
+      const context = { pathFormat: '/{urlKey}' };
+      expect(getProductUrl({ urlKey: 'my-url-key' }, context)).toBe(null);
+  });
+
+  test('return null for missing pathFormat', () => {
+      const context = { storeUrl: 'https://example.com' };
+      expect(getProductUrl({ urlKey: 'my-url-key' }, context)).toBe(null);
+  });
+
+  test('getProductUrl with path only', () => {
+      const context = { storeUrl: 'https://example.com', pathFormat: '/{locale}/products/{urlKey}/{sku}', locale: 'de' };
+      expect(getProductUrl({ urlKey: 'my-url-key', sku: 'my-sku' }, context, false)).toBe('/de/products/my-url-key/my-sku');
   });
 });
